third_party/blink/web_tests/animations/animationworklet/peek-updated-composited-property-on-main-expected.txt

CONSOLE MESSAGE: opacity of target is 0.5
